pyproject.toml
[tool.poetry]
name = "agents-pydanticai-fastapi"
version = "0.1.0"
description = ""
authors = ["Shailesh <shailesh1294p@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.1.0"
colorama = "^0.4.6"
pydantic-ai = "^0.1.3"
groq = "^0.22.0"
black = "^25.1.0"
fastapi = {extras = ["standard"], version = "^0.115.12"}

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"
